Reverting AzkabanDslCompiler update that is relevant to future PR
